feat(settings): extract settings interfaces
