+++ server/lib/settings/index.ts
@@ -1,7 +1,23 @@
 import { MediaServerType } from '@server/constants/server';
+import {
+  MetadataProviderType,
+  type AllSettings,
+  type FullPublicSettings,
+  type JellyfinSettings,
+  type JobId,
+  type JobSettings,
+  type MainSettings,
+  type MetadataSettings,
+  type NetworkSettings,
+  type NotificationSettings,
+  type PlexSettings,
+  type PublicSettings,
+  type RadarrSettings,
+  type SonarrSettings,
+  type TautulliSettings,
+} from '@server/interfaces/settings';
 import { Permission } from '@server/lib/permissions';
 import { runMigrations } from '@server/lib/settings/migrator';
-import type { AvailableLocale } from '@server/types/languages';
 import { randomBytes, randomUUID } from 'crypto';
 import fs from 'fs/promises';
 import { mergeWith } from 'lodash';
@@ -13,382 +29,6 @@ const mergeSettings = <T>(current: T, incoming: Partial<T>): T =>
   mergeWith({}, current, incoming, (_objValue, srcValue) =>
     Array.isArray(srcValue) ? srcValue : undefined
   ) as T;
-
-export interface Library {
-  id: string;
-  name: string;
-  enabled: boolean;
-  type: 'show' | 'movie';
-  lastScan?: number;
-}
-
-export interface Region {
-  iso_3166_1: string;
-  english_name: string;
-  name?: string;
-}
-
-export interface Language {
-  iso_639_1: string;
-  english_name: string;
-  name: string;
-}
-
-export interface PlexSettings {
-  name: string;
-  machineId?: string;
-  ip: string;
-  port: number;
-  useSsl?: boolean;
-  libraries: Library[];
-  webAppUrl?: string;
-}
-
-export interface JellyfinSettings {
-  name: string;
-  ip: string;
-  port: number;
-  useSsl?: boolean;
-  urlBase?: string;
-  externalHostname?: string;
-  jellyfinForgotPasswordUrl?: string;
-  libraries: Library[];
-  serverId: string;
-  apiKey: string;
-}
-export interface TautulliSettings {
-  hostname?: string;
-  port?: number;
-  useSsl?: boolean;
-  urlBase?: string;
-  apiKey?: string;
-  externalUrl?: string;
-}
-
-export interface DVRSettings {
-  id: number;
-  name: string;
-  hostname: string;
-  port: number;
-  apiKey: string;
-  useSsl: boolean;
-  baseUrl?: string;
-  activeProfileId: number;
-  activeProfileName: string;
-  activeDirectory: string;
-  tags: number[];
-  is4k: boolean;
-  isDefault: boolean;
-  externalUrl?: string;
-  syncEnabled: boolean;
-  preventSearch: boolean;
-  tagRequests: boolean;
-  overrideRule: number[];
-}
-
-export interface RadarrSettings extends DVRSettings {
-  minimumAvailability: string;
-}
-
-export interface SonarrSettings extends DVRSettings {
-  seriesType: 'standard' | 'daily' | 'anime';
-  animeSeriesType: 'standard' | 'daily' | 'anime';
-  activeAnimeProfileId?: number;
-  activeAnimeProfileName?: string;
-  activeAnimeDirectory?: string;
-  activeAnimeLanguageProfileId?: number;
-  activeLanguageProfileId?: number;
-  animeTags?: number[];
-  enableSeasonFolders: boolean;
-  monitorNewItems: 'all' | 'none';
-}
-
-interface Quota {
-  quotaLimit?: number;
-  quotaDays?: number;
-}
-
-export enum MetadataProviderType {
-  TMDB = 'tmdb',
-  TVDB = 'tvdb',
-}
-
-export interface MetadataSettings {
-  tv: MetadataProviderType;
-  anime: MetadataProviderType;
-}
-
-export interface ProxySettings {
-  enabled: boolean;
-  hostname: string;
-  port: number;
-  useSsl: boolean;
-  user: string;
-  password: string;
-  bypassFilter: string;
-  bypassLocalAddresses: boolean;
-}
-
-export interface MainSettings {
-  apiKey: string;
-  applicationTitle: string;
-  applicationUrl: string;
-  cacheImages: boolean;
-  defaultPermissions: number;
-  defaultQuotas: {
-    movie: Quota;
-    tv: Quota;
-  };
-  hideAvailable: boolean;
-  hideBlocklisted: boolean;
-  localLogin: boolean;
-  mediaServerLogin: boolean;
-  newPlexLogin: boolean;
-  discoverRegion: string;
-  streamingRegion: string;
-  originalLanguage: string;
-  blocklistRegion: string;
-  blocklistLanguage: string;
-  blocklistedTags: string;
-  blocklistedTagsLimit: number;
-  mediaServerType: number;
-  partialRequestsEnabled: boolean;
-  enableSpecialEpisodes: boolean;
-  locale: string;
-  youtubeUrl: string;
-  versionCheck: boolean;
-}
-
-export interface ProxySettings {
-  enabled: boolean;
-  hostname: string;
-  port: number;
-  useSsl: boolean;
-  user: string;
-  password: string;
-  bypassFilter: string;
-  bypassLocalAddresses: boolean;
-}
-
-export interface DnsCacheSettings {
-  enabled: boolean;
-  forceMinTtl?: number;
-  forceMaxTtl?: number;
-}
-
-export interface NetworkSettings {
-  csrfProtection: boolean;
-  forceIpv4First: boolean;
-  trustProxy: boolean;
-  proxy: ProxySettings;
-  dnsCache: DnsCacheSettings;
-  apiRequestTimeout: number;
-}
-
-interface PublicSettings {
-  initialized: boolean;
-}
-
-interface FullPublicSettings extends PublicSettings {
-  applicationTitle: string;
-  applicationUrl: string;
-  hideAvailable: boolean;
-  hideBlocklisted: boolean;
-  localLogin: boolean;
-  mediaServerLogin: boolean;
-  movie4kEnabled: boolean;
-  series4kEnabled: boolean;
-  discoverRegion: string;
-  streamingRegion: string;
-  originalLanguage: string;
-  mediaServerType: number;
-  jellyfinExternalHost?: string;
-  jellyfinForgotPasswordUrl?: string;
-  jellyfinServerName?: string;
-  partialRequestsEnabled: boolean;
-  enableSpecialEpisodes: boolean;
-  cacheImages: boolean;
-  vapidPublic: string;
-  enablePushRegistration: boolean;
-  locale: string;
-  emailEnabled: boolean;
-  userEmailRequired: boolean;
-  newPlexLogin: boolean;
-  youtubeUrl: string;
-  versionCheck: boolean;
-  plexClientIdentifier: string;
-}
-
-export interface NotificationAgentConfig {
-  enabled: boolean;
-  embedPoster: boolean;
-  types?: number;
-  options: Record<string, unknown>;
-}
-export interface NotificationAgentDiscord extends NotificationAgentConfig {
-  options: {
-    botUsername?: string;
-    botAvatarUrl?: string;
-    webhookUrl: string;
-    webhookRoleId?: string;
-    webhookThreadId?: string;
-    enableMentions: boolean;
-    locale: AvailableLocale;
-    useUserLocale: boolean;
-  };
-}
-
-export interface NotificationAgentSlack extends NotificationAgentConfig {
-  options: {
-    webhookUrl: string;
-    locale: AvailableLocale;
-  };
-}
-
-export interface NotificationAgentEmail extends NotificationAgentConfig {
-  options: {
-    userEmailRequired: boolean;
-    emailFrom: string;
-    smtpHost: string;
-    smtpPort: number;
-    secure: boolean;
-    ignoreTls: boolean;
-    requireTls: boolean;
-    authUser?: string;
-    authPass?: string;
-    allowSelfSigned: boolean;
-    senderName: string;
-    usePublicLogo: boolean;
-    pgpPrivateKey?: string;
-    pgpPassword?: string;
-  };
-}
-
-export interface NotificationAgentTelegram extends NotificationAgentConfig {
-  options: {
-    botUsername?: string;
-    botAPI: string;
-    chatId: string;
-    messageThreadId: string;
-    sendSilently: boolean;
-  };
-}
-
-export interface NotificationAgentPushbullet extends NotificationAgentConfig {
-  options: {
-    accessToken: string;
-    channelTag?: string;
-  };
-}
-
-export interface NotificationAgentPushover extends NotificationAgentConfig {
-  options: {
-    accessToken: string;
-    userToken: string;
-    sound: string;
-  };
-}
-
-export interface NotificationAgentWebhook extends NotificationAgentConfig {
-  options: {
-    webhookUrl: string;
-    jsonPayload: string;
-    authHeader?: string;
-    customHeaders?: { key: string; value: string }[];
-    supportVariables?: boolean;
-  };
-}
-
-export interface NotificationAgentGotify extends NotificationAgentConfig {
-  options: {
-    url: string;
-    token: string;
-    priority: number;
-    locale: AvailableLocale;
-  };
-}
-
-export interface NotificationAgentNtfy extends NotificationAgentConfig {
-  options: {
-    url: string;
-    topic: string;
-    authMethodUsernamePassword?: boolean;
-    username?: string;
-    password?: string;
-    authMethodToken?: boolean;
-    token?: string;
-    priority?: number;
-    locale: AvailableLocale;
-  };
-}
-
-export enum NotificationAgentKey {
-  DISCORD = 'discord',
-  EMAIL = 'email',
-  GOTIFY = 'gotify',
-  NTFY = 'ntfy',
-  PUSHBULLET = 'pushbullet',
-  PUSHOVER = 'pushover',
-  SLACK = 'slack',
-  TELEGRAM = 'telegram',
-  WEBHOOK = 'webhook',
-  WEBPUSH = 'webpush',
-}
-
-interface NotificationAgents {
-  discord: NotificationAgentDiscord;
-  email: NotificationAgentEmail;
-  gotify: NotificationAgentGotify;
-  ntfy: NotificationAgentNtfy;
-  pushbullet: NotificationAgentPushbullet;
-  pushover: NotificationAgentPushover;
-  slack: NotificationAgentSlack;
-  telegram: NotificationAgentTelegram;
-  webhook: NotificationAgentWebhook;
-  webpush: NotificationAgentConfig;
-}
-
-interface NotificationSettings {
-  agents: NotificationAgents;
-}
-
-interface JobSettings {
-  schedule: string;
-}
-
-export type JobId =
-  | 'plex-recently-added-scan'
-  | 'plex-full-scan'
-  | 'plex-watchlist-sync'
-  | 'plex-refresh-token'
-  | 'radarr-scan'
-  | 'sonarr-scan'
-  | 'download-sync'
-  | 'download-sync-reset'
-  | 'jellyfin-recently-added-scan'
-  | 'jellyfin-full-scan'
-  | 'image-cache-cleanup'
-  | 'availability-sync'
-  | 'process-blocklisted-tags';
-
-export interface AllSettings {
-  clientId: string;
-  sessionSecret?: string;
-  vapidPublic: string;
-  vapidPrivate: string;
-  main: MainSettings;
-  plex: PlexSettings;
-  jellyfin: JellyfinSettings;
-  tautulli: TautulliSettings;
-  radarr: RadarrSettings[];
-  sonarr: SonarrSettings[];
-  public: PublicSettings;
-  notifications: NotificationSettings;
-  jobs: Record<JobId, JobSettings>;
-  network: NetworkSettings;
-  metadataSettings: MetadataSettings;
-  migrations: string[];
-}
 
 const SETTINGS_PATH = process.env.CONFIG_DIRECTORY
   ? `${process.env.CONFIG_DIRECTORY}/settings.json`
